IllegalRoleNameException moved to exception.roles package, AccessDeniedException moved to exception.security package
