--- src/main/java/org/blbulyandavbulyan/blog/exceptions/AccessDeniedException.java
@@ -1,9 +1,0 @@
-package org.blbulyandavbulyan.blog.exceptions;
-
-import org.springframework.http.HttpStatus;
-
-public class AccessDeniedException extends BlogException{
-    public AccessDeniedException() {
-        super("Operation not permitted", HttpStatus.FORBIDDEN);
-    }
-}
