+++ src/main/java/org/blbulyandavbulyan/blog/services/ArticlesService.java
@@ -7,8 +7,8 @@ import org.blbulyandavbulyan.blog.dtos.article.ArticleResponse;
 import org.blbulyandavbulyan.blog.dtos.article.CreateArticleRequest;
 import org.blbulyandavbulyan.blog.entities.Article;
 import org.blbulyandavbulyan.blog.entities.User;
-import org.blbulyandavbulyan.blog.exceptions.AccessDeniedException;
 import org.blbulyandavbulyan.blog.exceptions.articles.ArticleNotFoundException;
+import org.blbulyandavbulyan.blog.exceptions.security.AccessDeniedException;
 import org.blbulyandavbulyan.blog.repositories.ArticleRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
