+++ src/test/java/org/blbulyandavbulyan/blog/services/RoleServiceTest.java
@@ -1,7 +1,7 @@
 package org.blbulyandavbulyan.blog.services;
 
 import org.blbulyandavbulyan.blog.entities.Role;
-import org.blbulyandavbulyan.blog.exceptions.IllegalRoleNameException;
+import org.blbulyandavbulyan.blog.exceptions.role.IllegalRoleNameException;
 import org.blbulyandavbulyan.blog.repositories.RoleRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
